refactor: renaming js files using camel case
