+++ src/components/LoggedNav.js
@@ -1,9 +1,10 @@
 import React, {useState} from 'react';
-import HomeView from '../views/home-view';
-import ProfileView from '../views/profile-view';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {UserContext} from '../context/user-context';
-import getProfile from '../services/get-profile';
+
+import HomeView from '../views/HomeView';
+import ProfileView from '../views/ProfileView';
+import getProfile from '../services/GetProfile';
+import {UserContext} from '../context/UserContext';
 
 const Stack = createNativeStackNavigator();
 const DEFAULT_URL_USER_PICTURE = 'https://cdn.icon-icons.com/icons2/3065/PNG/512/profile_user_account_icon_190938.png';
