+++ src/components/UnloggedNav.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import SignUpView from '../views/sign-up-view';
-import LoginView from '../views/login-view';
-import ResetPasswordView from '../views/reset-password';
+
+import SignUpView from '../views/SignUpView';
+import LoginView from '../views/LoginView';
+import ResetPasswordView from '../views/ResetPassword';
 
 const Stack = createNativeStackNavigator();
