+++ src/views/LoginView.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import {View} from 'react-native';
-import SignUpForm from '../components/sign-up-form';
 
-export default function SignUpView(props) {
+import LoginForm from '../components/LoginForm';
+
+export default function LoginView(props) {
   return (
     <View style={{flex: 1}}>
-      <SignUpForm
+      <LoginForm
+        onLogin={props.onLogin}
         onNavigation={props.onNavigation.navigation}
       />
     </View>
