+++ src/views/SignUpView.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import {View} from 'react-native';
-import Home from '../components/home';
 
-export default function HomeView(props) {
+import SignUpForm from '../components/SignUpForm';
+
+export default function SignUpView(props) {
   return (
     <View style={{flex: 1}}>
-      <Home
+      <SignUpForm
         onNavigation={props.onNavigation.navigation}
       />
     </View>
   );
 }
-
